notification in progress

+++ CPO Swap/ViewModels/ShiftDetailViewModel.swift
@@ -9,12 +9,19 @@ import Foundation
 import FirebaseFirestore
 
 class ShiftDetailViewModel  : ObservableObject {
-    init(){
+    @Published var enteredReg : String = ""
+    @Published var enteredLast : String = ""
+    @Published var DeleteerrorMessage : String = ""
+    
+    func delete(uid : String, reg: String, last: String){
+        let db = Firestore.firestore()
         
+        if enteredReg == reg && enteredLast == last {
+            db.collection("shifts").document(uid).delete()
+        } else {
+            DeleteerrorMessage = "Entered Wrong Reg# or Last Name"
+        }
     }
     
-    func delete(uid : String){
-        let db = Firestore.firestore()
-        db.collection("shifts").document(uid).delete()
-    }
+    
 }
